db/requirements.txt
Flask==0.10.1
requests==2.4.3
docker-py==1.0.0
PyMySQL==0.6.4
gunicorn==19.3.0
